+++ implementations/cpp/algorithms/concurrent/decompressor.cpp
@@ -3,6 +3,7 @@
 #include <algorithm>
 
 #include "../../fractal/block/reduction/reduction-strategy.hpp"
+#include "../../fractal/transformation/transformation.hpp"
 #include "../../utils/matrix.hpp"
 
 Decompressor::Decompressor(int domainSize, int rangeSize, ReductionStrategy* reductionStrategy)
@@ -18,10 +19,15 @@ void Decompressor::apply(const FractalMapping& mapping, Matrix<GrayPixel>& img, 
 
     float s = mapping.s;
     float o = mapping.o;
+    TransformationType transformation_type = mapping.transformation_type;
+    Transformation* transformation = Transformation::from(transformation_type);
+    Matrix<GrayPixel> transformedDomain;
+    transformedDomain.resize(reducedDomain->getRows(), reducedDomain->getCols());
+    transformation->transform(reducedDomain, &transformedDomain);
 
     for (int i = 0; i < reducedDomain->getRows(); i++) {
         for (int j = 0; j < reducedDomain->getCols(); j++) {
-            GrayPixel gp = reducedDomain->get(i, j);
+            GrayPixel gp = transformedDomain.get(i, j);
 
             int newLevel = static_cast<int>(s * static_cast<float>(gp.level) + o);
             newLevel = std::max(0, std::min(255, newLevel));