+++ implementations/cpp/fractal/block/compressed.cpp
@@ -1,6 +1,6 @@
 #include "compressed.hpp"
 
-CompressedBlock::CompressedBlock() : range(nullptr), domain(nullptr), s(0), o(0) {}
+CompressedBlock::CompressedBlock() : range(nullptr), domain(nullptr), s(0), o(0), transformation_type(TransformationType::IDENTITY) {}
 
-CompressedBlock::CompressedBlock(Block* range, Block* domain, float s, float o)
-    : range(range), domain(domain), s(s), o(o) {}
+CompressedBlock::CompressedBlock(Block* range, Block* domain, float s, float o, TransformationType transformation_type)
+    : range(range), domain(domain), s(s), o(o), transformation_type(transformation_type) {}