+++ implementations/cpp/fractal/mapping/fractal-mapping.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-
+#include "../transformation/type.hpp"
 
 class FractalMapping{
     
@@ -11,6 +11,7 @@ class FractalMapping{
         int domain_y;
         float s;
         float o;
+        TransformationType transformation_type;
         FractalMapping();
-        FractalMapping(int range_x, int range_y, int domain_x, int domain_y, float s, float o);
+        FractalMapping(int range_x, int range_y, int domain_x, int domain_y, float s, float o, TransformationType transformation_type);
 };